Move all JSON serializers to blueflood-http
