+++ blueflood-http/src/main/java/com/rackspacecloud/blueflood/outputs/serializers/BatchedMetricsJSONOutputSerializer.java
