+++ blueflood-http/src/test/java/com/rackspacecloud/blueflood/outputs/serializers/BatchedMetricsJSONOutputSerializerTest.java
